Add files via upload
Hate speech detection (HSD) is used to identify communication that contains hatred or encourages violence against a person or group of people. It can be applied to text, audio, and other forms of communication. HSD is often based on prejudice against protected characteristics such as gender, ethnicity, religion, sexual orientation, and age.